feat: working example on iOS
