+++ ios/observers/FocusedInputObserver.swift
@@ -25,10 +25,13 @@ let noFocusedInputEvent: [String: Any] = [
 @objc(FocusedInputObserver)
 public class FocusedInputObserver: NSObject {
   // class members
+  /// handlers
   var onLayoutChangedHandler: (NSDictionary) -> Void
   var onTextChangedHandler: (String) -> Void
   var onSelectionChangedHandler: (NSDictionary) -> Void
   var onFocusDidSet: (NSDictionary) -> Void
+  /// delegates
+  var delegate: KCTextInputCompositeDelegate
   // state variables
   private var isMounted = false
   // input tracking
@@ -48,6 +51,8 @@ public class FocusedInputObserver: NSObject {
     self.onTextChangedHandler = onTextChangedHandler
     self.onSelectionChangedHandler = onSelectionChangedHandler
     self.onFocusDidSet = onFocusDidSet
+      // TODO: weak reference
+    self.delegate = KCTextInputCompositeDelegate(onSelectionChange: self.onSelectionChangedHandler)
   }
 
   @objc public func mount() {
@@ -128,22 +133,6 @@ public class FocusedInputObserver: NSObject {
 
     if let string = text {
       onTextChangedHandler(string)
-        onSelectionChangedHandler([
-            "position": [
-                "start": 0,
-                "end": 12
-            ],
-            "coordinates": [
-              "start": [
-                "x": 0,
-                "y": 50
-              ],
-              "end": [
-                "x": 20,
-                "y": 100
-              ]
-            ],
-          ])
     }
   }
 
@@ -163,8 +152,11 @@ public class FocusedInputObserver: NSObject {
 
     if currentInput != nil {
       hasObservers = true
+      let responder = UIResponder.current
       currentInput?.addObserver(self, forKeyPath: "center", options: .new, context: nil)
-      textChangeObserver.observeTextChanges(for: UIResponder.current, handler: onTextChanged)
+      textChangeObserver.observeTextChanges(for: responder, handler: onTextChanged)
+      
+        substituteDelegate(responder)
     }
   }
 
@@ -177,6 +169,21 @@ public class FocusedInputObserver: NSObject {
     currentInput?.removeObserver(self, forKeyPath: "center", context: nil)
     textChangeObserver.removeObserver()
   }
+    
+    private func substituteDelegate(_ input: UIResponder?) {
+        if let textField = input as? UITextField {
+            if (!(textField.delegate is KCTextInputCompositeDelegate)) {
+                delegate.setTextFieldDelegate(delegate: textField.delegate)
+                textField.delegate = delegate
+            }
+        } else if let textView = input as? UITextView {
+            if (!(textView.delegate is KCTextInputCompositeDelegate)) {
+                delegate.setTextViewDelegate(delegate: textView.delegate)
+                (textView as? RCTUITextView)?.setForceDelegate(delegate)
+            }
+            print(textView.delegate)
+        }
+    }
 
   // swiftlint:disable:next block_based_kvo
   @objc override public func observeValue(
